This code adds entity framework support for the app's order and pizza classes.
In Visual Studio Code, on the menu, select File > New Text File.

Select the C# language, and enter this code:

The preceding code allows our app to get all the current orders and place an order. The [Route("orders")] Blazor attribute allows this class to handle incoming HTTP requests for /orders and /orders/{orderId}.

Save your changes with Ctrl+S.

For the filename, use OrderController.cs. Make sure you save the file in the same directory as OrderState.cs.

In the file explorer, select OrderState.cs.

At the bottom of the class under the RemoveConfiguredPizza method, modify ResetOrder() to reset the order: