In the file explorer, expand Pages, right-click the folder, and select New File.
Name the new file MyOrders.razor. In this file, write the following code:

The navigation needs to change on all the pages we have now to include a link to the new My orders page. Open Checkout.razor and Index.razor and replace the navigation with this code:

By using <a> elements, we can manage which is the active page manually by adding the active css class. Let's update all the navigation to use a NavLink component instead.