In the file explorer, select OrderController.cs.
Under the PlaceOrder method, add a new method to return orders with a status.
This code enabled the Order controller to respond to an HTTP request with orderId in the URL. The method then uses this ID to query the database and, if an order is found, return an OrderWithStatus object.

Let's use this new page when a customer checks out. You need to update the Checkout.razor component.