In Visual Studio Code, on the menu, select File > New Text File.
Select ASP.NET Razor as the language.

Create a layout component, and copy the navigation HTML from any page.
This layout has used some of the markup in _Host.cshtml, so we need to remove it from there.

Select Ctrl+S to save your changes.

For the filename, use MainLayout.razor. Make sure you save the file in the Shared directory.