In Visual Studio Code, in the file explorer, expand Pages, then select Checkout.razor.
At the moment, the checkout page just shows the list of pizzas that have been configured. You've been asked to refactor the page to include an address section.

Replace the existing <div class="main"> HTML block to have two columns for data.

The preceding code connects the HTML onclick event for the button element to the PlaceOrder Blazor method in the @code block.

The refactoring references two new Blazor controls OrderReview and AddressEditor. You'll move the old checkout code that listed the pizzas to the OrderReview component.

In the file explorer, right-click on Shared, then select New File.

Enter OrderReview.razor as the filename.

Add a foreach loop to display the pizzas in an order.