Initialize the context with the order delivery address.
This code also allows for linking an event handler to when a field is changed. In the new handler, you can check to see if the model is valid and set the isError appropriately.

Because we've now created an event handler, we should dispose of it when the checkout component no longer needs it.

To implement the Dispose feature, you also have to let Blazor know. Add this code at the top of the page under the @inject statements.

Delete all the references to isSubmitting and update the PlaceOrder method.